backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
psycopg[binary]==3.2.9
pydantic==2.11.7
pydantic-settings==2.10.1
email-validator==2.2.0
python-multipart==0.0.20
google-generativeai==0.8.5
